+++ Source/SelfService/Web/microservice/microserviceView/configuration/editSection.tsx
@@ -18,7 +18,7 @@ import { canDeleteMicroservice, deleteMicroservice, MicroserviceStore } from '..
 import { HttpResponseApplication } from '../../../api/application';
 
 import { HeadArguments } from '../../components/headArguments';
-import { RestartMicroserviceDialog } from '../RestartMicroserviceDialog';
+import { RestartMicroserviceDialog } from '../restartMicroserviceDialog';
 import { capitalize, removeFromString } from '../../helpers/stringHelpers';
 
 const styles = {
