+++ Source/SelfService/Web/microservice/overview.tsx
@@ -27,6 +27,8 @@ export const Overview = ({ application, microserviceId, environment }: OverviewP
 
     const [loaded, setLoaded] = useState(false);
 
+    console.log('Overview', microserviceId)
+
     // Want microservice name
     const [podsData, setPodsData] = useState({
         namespace: '',
